% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hOmega.R
\name{hOmega}
\alias{hOmega}
\title{Solution path function for the Penalized Interaction Estimation (PIE)}
\usage{
hOmega(X, Y, lambda = NULL, nlambda = 50, lambda.min.ratio = 0.01,
  err = 10^(-4), maxIter = 10^3, rho = 1)
}
\arguments{
\item{X}{covariates data matrix of dimension n*p after centering.}

\item{Y}{response variable with length n.}

\item{lambda}{user supplied tuning parameter;}

\item{err}{the precision used to control the convergence. Default is 1e-4.}

\item{maxIter}{Maximum number of iterations. Default is 1000.}

\item{rho}{step parameter for the ADMM. Default is 1.}
}
\value{
A list of sparse asymmetric p*p matrices, lambda and also the number of iterations.
}
\description{
Solution path function for the Penalized Interaction Estimation (PIE)
}
